+++ htmldoc/zipc.h
@@ -3,7 +3,7 @@
  *
  *     https://github.com/michaelrsweet/zipc
  *
- * Copyright 2017 by Michael R Sweet.
+ * Copyright 2017-2018 by Michael R Sweet.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -42,6 +42,10 @@
  * Types...
  */
 
+#  ifdef WIN32
+typedef long ssize_t;
+#  endif /* WIN32 */
+
 typedef struct _zipc_s zipc_t;		/* ZIP container */
 typedef struct _zipc_file_s zipc_file_t;/* File/directory in ZIP container */
 
@@ -61,19 +65,24 @@ extern zipc_file_t	*zipcCreateFile(zipc_t *zc, const char *filename, int compres
 extern int		zipcCreateFileWithString(zipc_t *zc, const char *filename, const char *contents);
 extern const char	*zipcError(zipc_t *zc);
 extern int		zipcFileFinish(zipc_file_t *zf);
+extern int              zipcFileGets(zipc_file_t *zf, char *line, size_t linesize);
 extern int		zipcFilePrintf(zipc_file_t *zf, const char *format, ...)
 #  ifdef __GNUC__
 __attribute__ ((__format__ (__printf__, 2, 3)))
 #  endif /* __GNUC__ */
 ;
 extern int		zipcFilePuts(zipc_file_t *zf, const char *s);
+extern ssize_t		zipcFileRead(zipc_file_t *zf, void *data, size_t bytes);
 extern int		zipcFileWrite(zipc_file_t *zf, const void *data, size_t bytes);
+extern int              zipcFileXMLGets(zipc_file_t *zf, char *fragment, size_t fragsize);
 extern int		zipcFileXMLPrintf(zipc_file_t *zf, const char *format, ...)
 #  ifdef __GNUC__
 __attribute__ ((__format__ (__printf__, 2, 3)))
 #  endif /* __GNUC__ */
 ;
 extern zipc_t		*zipcOpen(const char *filename, const char *mode);
+extern zipc_file_t      *zipcOpenFile(zipc_t *zc, const char *filename);
+extern const char       *zipcXMLGetAttribute(const char *element, const char *attrname, char *buffer, size_t bufsize);
 
 #  ifdef __cplusplus
 }